+++ src/com/jumbo/entities/graphics/JumboButton.java
@@ -10,7 +10,7 @@ import com.jumbo.components.interfaces.TriggeredAction;
 import com.jumbo.components.interfaces.TriggeredEvent;
 import com.jumbo.core.JumboEntity;
 import com.jumbo.core.JumboGraphicsObject;
-import com.jumbo.core.JumboTexture;
+import com.jumbo.core.texture.JumboTexture;
 import com.jumbo.entities.graphics.text.JumboText;
 import com.jumbo.entities.graphics.text.JumboTextBox;
 import com.jumbo.tools.input.JumboInputListener;
@@ -418,6 +418,14 @@ public class JumboButton extends JumboGraphicsObject {
 		}
 	}
 
+	public void setDisabledIcon(JumboImage icon) {
+		this.disabledicon = icon;
+	}
+
+	public void setDisabledIcon(JumboTexture icon) {
+		this.disabledicon.setTexture(icon);
+	}
+
 	/**
 	 * @param hovericon
 	 *            the hovericon to set
@@ -452,6 +460,19 @@ public class JumboButton extends JumboGraphicsObject {
 		if (action != null) {
 			action.action();
 		}
+	}
+
+	/*
+	 * (non-Javadoc)
+	 * 
+	 * @see
+	 * com.jumbo.core.JumboEntity#additionalCalculations(java.awt.Rectangle)
+	 */
+	@Override
+	public Rectangle additionalCalculations(Rectangle inbounds) {
+		final Rectangle bounds = super.additionalCalculations(inbounds);
+
+		return bounds;
 	}
 
 }