+++ src/com/jumbo/entities/graphics/particles/JumboParticleField.java
@@ -7,7 +7,7 @@ import com.jumbo.components.JumboColor;
 import com.jumbo.components.MinMaxVector;
 import com.jumbo.core.JumboEntity;
 import com.jumbo.core.JumboGraphicsObject;
-import com.jumbo.core.JumboTexture;
+import com.jumbo.core.texture.JumboTexture;
 import com.jumbo.entities.JumboGraphicsGroup;
 import com.jumbo.entities.graphics.JumboAnimation;
 import com.jumbo.entities.graphics.JumboAnimationFrame;