+++ src/com/jumbo/entities/graphics/text/JumboInputBox.java
@@ -5,7 +5,7 @@ import java.awt.image.BufferedImage;
 import java.util.ArrayList;
 
 import com.jumbo.components.interfaces.TriggeredAction;
-import com.jumbo.core.JumboTexture;
+import com.jumbo.core.texture.JumboTexture;
 import com.jumbo.entities.graphics.JumboButton;
 import com.jumbo.entities.graphics.JumboImage;
 import com.jumbo.tools.input.JumboInputHandler;