+++ src/com/jumbo/entities/graphics/text/JumboText.java
@@ -5,7 +5,7 @@ import java.util.ArrayList;
 
 import com.jumbo.core.JumboEntity;
 import com.jumbo.core.JumboGraphicsObject;
-import com.jumbo.core.JumboTexture;
+import com.jumbo.core.texture.JumboTexture;
 import com.jumbo.entities.JumboGraphicsGroup;
 import com.jumbo.tools.loaders.JumboStringHandler;