+++ src/com/jumbo/tools/loaders/JumboStringHandler.java
@@ -15,7 +15,7 @@ import com.jumbo.components.FloatRectangle;
 import com.jumbo.components.JumboColor;
 import com.jumbo.core.Jumbo;
 import com.jumbo.core.JumboEntity;
-import com.jumbo.core.JumboTexture;
+import com.jumbo.core.texture.JumboTexture;
 import com.jumbo.entities.graphics.text.JumboLetter;
 import com.jumbo.tools.JumboErrorHandler;
 import com.jumbo.tools.JumboSettings;